pyproject.toml
[tool.poetry]
name = "carrier-pigeon"
version = "0.1.0"
description = ""
authors = ["Daniel Morrison <dmorrison@boxboat.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
asyncio-nats-client = "^0.11.4"
Jinja2 = "^3.0.2"
python-multipart = "^0.0.5"
uvicorn = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
